+++ src/main/php/CallAmountViolation.php
@@ -6,13 +6,13 @@
  * file that was distributed with this source code.
  */
 namespace bovigo\callmap;
-if (class_exists('PHPUnit_Framework_ExpectationFailedException')) {
+if (class_exists('PHPUnit\Framework\ExpectationFailedException')) {
     /**
      * Exception to be thrown when the call amount of a method is not the expected amount of times.
      *
      * @since  0.5.0
      */
-    class CallAmountViolation extends \PHPUnit_Framework_ExpectationFailedException
+    class CallAmountViolation extends \PHPUnit\Framework\ExpectationFailedException
     {
         // intentionally empty
     }
@@ -25,4 +25,3 @@ if (class_exists('PHPUnit_Framework_ExpectationFailedException')) {
         // intentionally empty
     }
 }
-