+++ src/main/php/functions.php
@@ -154,9 +154,9 @@ namespace bovigo\callmap {
     /**
      * blacklist our own classes from being displayed in PHPUnit error stacks
      */
-    if (class_exists('\PHPUnit_Util_Blacklist')) {
-        \PHPUnit_Util_Blacklist::$blacklistedClassNames = array_merge(
-                \PHPUnit_Util_Blacklist::$blacklistedClassNames,
+    if (class_exists('\PHPUnit\Util\Blacklist')) {
+        \PHPUnit\Util\Blacklist::$blacklistedClassNames = array_merge(
+                \PHPUnit\Util\Blacklist::$blacklistedClassNames,
                 [Verification::class => 1]
         );
     }